break down SourcePointRequestsResponses into multiple files
